Rename 'test-parser' to 'parser-types'
