+++ src/main.ts
@@ -2,7 +2,7 @@ import * as core from '@actions/core'
 import * as github from '@actions/github'
 import {parseJestJunit} from './parsers/jest-junit/jest-junit-parser'
 import {parseDartJson} from './parsers/dart-json/dart-json-parser'
-import {ParseOptions, ParseTestResult} from './parsers/test-parser'
+import {ParseOptions, ParseTestResult} from './parsers/parser-types'
 import {getFileContent, normalizeDirPath} from './utils/file-utils'
 import {listFiles} from './utils/git'
 import {getCheckRunSha} from './utils/github-utils'
